requirements.txt
fastapi
uvicorn
cassandra-driver
python-dotenv
pydantic-settings
email-validator
argon2-cffi
jupyter
pytest
jinja2
python-multipart
python-jose[cryptography]
algoliasearch